chore: move State to own file
