+++ AppCore/Sources/Screens/Favorites/Store/FavoritesPresenter.swift
@@ -14,43 +14,6 @@ import Helpers
 @MainActor
 public final class FavoritesPresenter {
 
-    struct State: Equatable {
-
-        // MARK: - Properties
-
-        var filteredRecipes: [Recipe]
-        var pendingRecipe: Recipe?
-        var query: String
-        var recipes: DomainModelState<[Recipe]>
-        var route: AnyIdentifiable<Route>?
-
-        var areFavoriteRecipesEmpty: Bool {
-            recipes.isPresent && recipes.isEmpty && recipes.error == nil
-        }
-
-        var isError: Bool {
-            recipes.isEmpty && recipes.error != nil
-        }
-
-        // MARK: - Public methods
-
-        mutating func updateFilteredRecipes() {
-            guard !query.isEmpty else {
-                self.filteredRecipes = recipes.items
-                return
-            }
-
-            let lowercasedQuery = query.lowercased()
-            self.filteredRecipes = recipes.items.filter { $0.name.lowercased().contains(lowercasedQuery) }
-        }
-    }
-
-    enum Route {
-        case didTapExplore
-        case didTapFilters
-        case didTapRecipe(Recipe)
-    }
-
     // MARK: - Properties
 
     @Dependency
@@ -63,7 +26,7 @@ public final class FavoritesPresenter {
     // MARK: - Lifecycle
 
     public init() {
-        self.state = State(filteredRecipes: [], query: "", recipes: DomainModelState(), route: nil)
+        self.state = State.makeInitialState()
         Task {
             await observeRecipes()
         }
